refactor: 大規模ファイル5件を分割
## 分割対象と結果

### 1. UnifiedDrawingCanvas.tsx (930行 → 212行)
- utils/coordinate-utils.ts: 座標変換関数
- utils/line-rendering.ts: 線描画ロジック
- utils/hit-detection.ts: 当たり判定
- utils/canvas-renderer.ts: 再描画エンジン
- hooks/useCanvasMouseEvents.ts: マウスイベント処理

### 2. ScoringMainView.tsx (836行 → 409行)
- hooks/useScoringShortcuts.ts: キーボードショートカット
- hooks/useScoringEffects.ts: Effect処理
- hooks/useStudentAnswerManagement.ts: 学生管理
- hooks/useScoringActions.ts: 採点アクション

### 3. AnswerIndividualView.tsx (950行 → 492行)
- hooks/view/useZoomAndScroll.ts: ズーム・スクロール
- hooks/view/useDrawingToolShortcuts.ts: 描画ツールショートカット
- hooks/view/useQuestionAutoScroll.ts: 自動スクロール
- hooks/view/useCanvasV4Integration.ts: V4統合
- hooks/view/useTextInputStateNotifier.ts: テキスト入力通知
- hooks/view/useAllStudentAnnotations.ts: 全アノテーション取得

### 4. useImageCanvas.ts (1359行 → 106行)
- hooks/core/types.ts: 型定義
- hooks/core/useCanvasRefs.ts: キャンバス参照
- hooks/core/useImageLoader.ts: 画像読み込み
- hooks/core/useScoringMarks.ts: 採点記号
- hooks/core/useDrawingRenderer.ts: 描画レンダラー
- hooks/core/useCanvasDrawing.ts: Canvas描画統合

### 5. studentAnswer.ts (858行 → 26行)
- studentAnswer/crud.ts: CRUD操作
- studentAnswer/placement.ts: 配置管理
- studentAnswer/batch.ts: 一括操作
- studentAnswer/status.ts: ステータス管理
- studentAnswer/index.ts: 統合エクスポート

## 検証結果
- npm run lint: エラーなし
- npm run typecheck: エラーなし
