+++ components/drawing/utils/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @fileoverview 描画ユーティリティのエクスポート
+ */
+
+export { getRelativeCoordinates, getAbsoluteCoordinates } from "./coordinate-utils"
+export { drawWaveLine, drawZigzagLine, drawArrowHead } from "./line-rendering"
+export { isPointInAnnotation } from "./hit-detection"
+export { redrawCanvas, type CanvasDrawParams } from "./canvas-renderer"
